requirements.txt
south
django-custom-user
